move regex function to helpers file
